# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Opes360 API",
        "version": "1.0.0",
        "description": "Token-authenticated access to a business's records. A token can never do more than the person it belongs to could do while signed in.\n\nGenerated from the router by `php artisan opes:export-openapi` — do not edit by hand."
    },
    "servers": [
        {
            "url": "https://opes360.test"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
...
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/loyalty/contacts/{contact}/transactions": {
            "get": {
                "operationId": "loyaltyTransactions",
                "summary": "The points a customer has earned and spent.",
                "tags": [
                    "Loyalty"
                ],
                "parameters": [
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No token, or a revoked one"
                    },
                    "403": {
                        "description": "The permission, module switch or token scope says no"
                    },
                    "404": {
                        "description": "Not found, or belongs to another business"
                    },
                    "422": {
                        "description": "Validation failed, or a rule of the domain refused"
                    },
                    "429": {
                        "description": "Rate limited"
                    }
                },
                "description": "Requires the `read` token scope, on top of the permission the same action needs in the app.",
                "security": [
                    {
                        "bearerAuth": [
                            "read"
                        ]
                    }
                ]
            }
        },
...
    }
}